+++ src/main/java/com/sparklenote/student/controller/StudentController.java
@@ -1,0 +1,8 @@
+package com.sparklenote.student.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class StudentController {
+    // join/{url}/roll API에서 처리
+}
