+++ src/main/java/com/sparklenote/user/jwt/JWTFilter.java
@@ -3,8 +3,8 @@ package com.sparklenote.user.jwt;
 import com.sparklenote.domain.enumType.Role;
 import com.sparklenote.user.dto.response.UserResponseDTO;
 import com.sparklenote.user.oAuth2.CustomOAuth2User;
-import com.sparklenote.user.student.CustomStudentDetails;
-import com.sparklenote.user.student.dto.StudentResponseDTO;
+import com.sparklenote.student.userDetails.CustomStudentDetails;
+import com.sparklenote.student.dto.response.StudentResponseDTO;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
